part 2 wav files

+++ src/MP3Track.cpp
@@ -38,7 +38,7 @@ void MP3Track::analyze_beatgrid()
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
 
-    long beats = ((duration_seconds / 60.0) * bpm);
+    int beats = ((duration_seconds / 60.0) * bpm);
     double precision_factor = bitrate / 320.0;
     std::cout << "  -> Estimate beats : " << beats << "  → Compression precision factor:" << precision_factor << std::endl;
 }
